pyproject.toml
[project]
name = "evalmate"
version = "0.1.0"
description = "Intelligent Student Assignment Feedback System - Takes rubrics, assignment questions, and student submissions to generate structured feedback and grades using LLM evaluation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.4,<3",
    "pytest>=7.0.0",
    "pdfplumber>=0.10",
    "python-docx>=0.8",
    "camelot-py[cv]>=0.11.0",
    "tabula-py>=2.9.0",
    "regex>=2024.4.28",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "typer>=0.12",
    "python-multipart>=0.0.9",
    # Phase 6: Visual Extraction Dependencies
    "pillow>=10.2",
    "pytesseract>=0.3.10",
    "opencv-python-headless>=4.8",
    "pymupdf>=1.23",
    # Phase 7: Multimodal LLM Dependencies
    "openai>=1.3.0",
    "tenacity>=8.2.3",
    # Environment Configuration
    "python-dotenv>=1.0.0",
    "tiktoken>=0.12.0",
    "rich>=14.2.0",
    "pandas>=2.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests", "app/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | data
  | __pycache__
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["data/", "__pycache__/"]
